Split justfix-routes.tsx from justfix-site.tsx.
