+++ frontend/lib/data-driven-onboarding/data-driven-onboarding.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useContext } from "react";
 import classnames from "classnames";
-import JustfixRoutes from "../justfix-routes";
+import JustfixRoutes from "../justfix-route-info";
 import { RouteComponentProps } from "react-router";
 import Page, { PageTitle } from "../ui/page";
 import {
