+++ frontend/lib/data-requests/tests/routes.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import DataRequestsRoutes from "../routes";
-import JustfixRoutes from "../../justfix-routes";
+import JustfixRoutes from "../../justfix-route-info";
 import { DataRequestMultiLandlordQuery } from "../../queries/DataRequestMultiLandlordQuery";
 import { waitFor } from "@testing-library/react";
