+++ frontend/lib/dev/tests/routes.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import DevRoutes from "../routes";
-import JustfixRoutes from "../../justfix-routes";
+import JustfixRoutes from "../../justfix-route-info";
 
 describe("development pages", () => {
   it("shows development tools home", () => {
