+++ frontend/lib/global-site-routes.ts
@@ -1,7 +1,7 @@
 import { DevRouteInfo } from "./dev/route-info";
 import { RouteInfo, ROUTE_PREFIX } from "./util/route-util";
 import { getGlobalAppServerInfo, AppServerInfo } from "./app-context";
-import { default as JustfixRoutes } from "./justfix-routes";
+import { default as JustfixRoutes } from "./justfix-route-info";
 import { NorentRoutes } from "./norent/route-info";
 import History from "history";
