+++ frontend/lib/loc/tests/access-dates.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { getInitialState } from "../access-dates";
-import JustfixRoutes from "../../justfix-routes";
+import JustfixRoutes from "../../justfix-route-info";
 import LetterOfComplaintRoutes from "../routes";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import { AccessDatesMutation } from "../../queries/AccessDatesMutation";
