+++ frontend/lib/loc/tests/letter-request.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import JustfixRoutes from "../../justfix-routes";
+import JustfixRoutes from "../../justfix-route-info";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import LetterOfComplaintRoutes from "../routes";
 import { LetterRequestMutation } from "../../queries/LetterRequestMutation";
