+++ frontend/lib/loc/tests/loc-confirmation.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import JustfixRoutes from "../../justfix-routes";
+import JustfixRoutes from "../../justfix-route-info";
 import { LetterRequestMailChoice } from "../../queries/globalTypes";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import LetterOfComplaintRoutes from "../routes";
