+++ frontend/lib/pages/login-page.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 
 import Page from "../ui/page";
-import JustfixRoutes, { NEXT } from "../justfix-routes";
+import JustfixRoutes, { NEXT } from "../justfix-route-info";
 import { SessionUpdatingFormSubmitter } from "../forms/session-updating-form-submitter";
 import { LoginMutation, BlankLoginInput } from "../queries/LoginMutation";
 import { TextualFormField } from "../forms/form-fields";
