+++ frontend/lib/tests/util.tsx
@@ -7,7 +7,7 @@ import {
 } from "../app-context";
 import { AllSessionInfo, BlankAllSessionInfo } from "../queries/AllSessionInfo";
 import { FormError, strToFormError } from "../forms/form-errors";
-import JustfixRoutes from "../justfix-routes";
+import JustfixRoutes from "../justfix-route-info";
 import { Route } from "react-router-dom";
 
 interface TestClient {
